+++ lib/flutter_cache_manager_hive.dart
@@ -1,4 +1,4 @@
 library flutter_cache_manager_hive;
 
+export 'src/cache_object_adapter.dart';
 export 'src/hive_cache_manager.dart';
-export 'src/hive_storage/cache_object_adapter.dart';